Refine some of the previous changes to the auto updater
If we detect that initial navigation fails using the built-in
browser, we force delegation of all links to an external browser.

We also hide the Install/Download button and the UI related to the
location for the download/install package".  The message shown in the UI
instructs the user to click a link to download (which will then navigate
using an external browser)